Refactor MusclesChart component; rename and extract muscle rendering logic to a separate function for improved readability and maintainability
